requirements.txt
rembg
Pillow
PySimpleGUI
